refactor: API 내보내기 구조 개선 및 불필요한 파일 제거
