+++ src/entities/comment/index.ts
@@ -1,3 +1,7 @@
+// Model types
 export * from "./model/types"
 
-
+// API exports
+export { getComments } from "./api/services"
+export { commentQueryKeys } from "./api/query-keys"
+export { useGetComments } from "./api/hooks/use-get-comments"
