--- src/entities/post/api/index.ts
@@ -1,8 +1,0 @@
-// Services
-export { getPosts, searchPosts, getPostsByTag } from "./services"
-
-// Query Keys
-export { postQueryKeys } from "./query-keys"
-
-// Hooks
-export { useGetPostDetail } from "./hooks/use-get-post-detail"
