+++ src/entities/post/index.ts
@@ -1,16 +1,7 @@
-// Types
+// Model types
 export type { Post, PostsResponse, GetPostQuery, SearchPostQuery } from "./model/types"
 
-// API (Services, Queries, Mutations)
-export {
-  // Services
-  getPosts,
-  searchPosts,
-  getPostsByTag,
-
-  // Query Keys
-  postQueryKeys,
-
-  // Hooks
-  useGetPostDetail,
-} from "./api"
+// API exports
+export { getPosts, searchPosts, getPostsByTag } from "./api/services"
+export { postQueryKeys } from "./api/query-keys"
+export { useGetPostDetail } from "./api/hooks/use-get-post-detail"
