--- src/entities/tag/api/index.ts
@@ -1,5 +1,0 @@
-export { useGetTags } from "./hooks/use-get-tags"
-
-export { getTags } from "./services"
-
-export { tagQueryKeys } from "./query-keys"
