+++ src/entities/tag/index.ts
@@ -1,1 +1,7 @@
+// Model types
 export * from "./model/types"
+
+// API exports
+export { useGetTags } from "./api/hooks/use-get-tags"
+export { getTags } from "./api/services"
+export { tagQueryKeys } from "./api/query-keys"
