--- src/entities/user/api/index.ts
@@ -1,6 +1,0 @@
-export { useGetUsers } from "./hooks/use-get-users"
-export { useGetUserDetail } from "./hooks/use-get-user-detail"
-
-export { getUser, getUserById } from "./services"
-
-export { userQueryKeys } from "./query-keys"
