+++ src/entities/user/index.ts
@@ -1,1 +1,11 @@
+// Model types
 export * from "./model/types"
+
+// API exports
+export { useGetUsers } from "./api/hooks/use-get-users"
+export { useGetUserDetail } from "./api/hooks/use-get-user-detail"
+export { getUser, getUserById } from "./api/services"
+export { userQueryKeys } from "./api/query-keys"
+
+// UI exports
+export { UserDetailDialog } from "./ui/UserDetailDialog"
