+++ src/features/delete-comment/api/hooks/use-delete-comment-mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { CommentsResponse } from "../../../../entities/comment"
-import { commentQueryKeys } from "../../../../entities/comment/api/query-keys"
+import { commentQueryKeys } from "../../../../entities/comment"
 import { deleteComment } from "../services"
 
 export const useDeleteCommentMutation = () => {
