+++ src/features/edit-comment/api/hooks/use-update-comment-mutation.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { CommentsResponse } from "../../../../entities/comment"
-import { commentQueryKeys } from "../../../../entities/comment/api/query-keys"
+import { commentQueryKeys } from "../../../../entities/comment"
 import { updateComment, UpdateCommentRequestBody } from "../services"
 
 export const useUpdateCommentMutation = () => {
