+++ src/widgets/post-dashboard/ui/PostFilter.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Input, Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../../../shared/ui"
 import { Search } from "lucide-react"
-import { useGetTags } from "../../../entities/tag/api"
+import { useGetTags } from "../../../entities/tag"
 import { usePostListFilterSearchParams } from "../../../entities/post/model/hooks/use-post-list-filter-search-params"
 
 /**
