+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/index.tsx","./src/main.tsx","./src/app/app.tsx","./src/entities/comment/index.ts","./src/entities/comment/api/index.ts","./src/entities/comment/api/query-keys.ts","./src/entities/comment/api/services.ts","./src/entities/comment/api/hooks/use-get-comments.ts","./src/entities/comment/model/types.ts","./src/entities/post/index.ts","./src/entities/post/api/index.ts","./src/entities/post/api/query-keys.ts","./src/entities/post/api/services.ts","./src/entities/post/api/hooks/use-get-post-detail.ts","./src/entities/post/model/types.ts","./src/entities/tag/index.ts","./src/entities/tag/api/index.ts","./src/entities/tag/api/query-keys.ts","./src/entities/tag/api/services.ts","./src/entities/tag/api/hooks/use-get-tags.ts","./src/entities/tag/model/types.ts","./src/entities/user/index.ts","./src/entities/user/api/index.ts","./src/entities/user/api/query-keys.ts","./src/entities/user/api/services.ts","./src/entities/user/api/hooks/use-get-user-detail.ts","./src/entities/user/api/hooks/use-get-users.ts","./src/entities/user/model/types.ts","./src/features/comment/index.ts","./src/features/comment/api/services.ts","./src/features/comment/api/hooks/use-delete-comments-mutation.ts","./src/features/comment/api/hooks/use-like-comment-mutation.ts","./src/features/comment/api/hooks/use-post-comments-mutation.ts","./src/features/comment/api/hooks/use-update-comments-mutation.ts","./src/features/post/index.ts","./src/features/post/api/services.ts","./src/features/post/api/hooks/use-create-post-mutation.ts","./src/features/post/api/hooks/use-delete-post-mutation.ts","./src/features/post/api/hooks/use-update-post-mutation.ts","./src/pages/post/index.tsx","./src/pages/post/ui/commentssection.tsx","./src/pages/post/ui/index.ts","./src/shared/lib/axios-instance.ts","./src/shared/ui/index.ts","./src/shared/ui/button/button.tsx","./src/shared/ui/button/index.ts","./src/shared/ui/card/card.tsx","./src/shared/ui/card/index.ts","./src/shared/ui/dialog/dialog.tsx","./src/shared/ui/dialog/index.ts","./src/shared/ui/footer/footer.tsx","./src/shared/ui/footer/index.ts","./src/shared/ui/header/header.tsx","./src/shared/ui/header/index.ts","./src/shared/ui/input/input.tsx","./src/shared/ui/input/index.ts","./src/shared/ui/select/select.tsx","./src/shared/ui/select/index.ts","./src/shared/ui/table/table.tsx","./src/shared/ui/table/index.ts","./src/shared/ui/textarea/textarea.tsx","./src/shared/ui/textarea/index.ts","./src/widgets/post-list/api/hooks/use-get-post-list.ts","./src/widgets/post-list/model/post-list-query-schema.ts","./src/widgets/post-list/model/hooks/use-post-list-filter-search-params.ts","./src/widgets/post-list/ui/postlistexample.tsx","./src/widgets/post-list/ui/posttable.tsx","./vite-env.d.ts"],"errors":true,"version":"5.9.2"}
+{"root":["./src/index.tsx","./src/main.tsx","./src/app/app.tsx","./src/entities/comment/index.ts","./src/entities/comment/api/query-keys.ts","./src/entities/comment/api/services.ts","./src/entities/comment/api/hooks/use-get-comments.ts","./src/entities/comment/model/types.ts","./src/entities/post/index.ts","./src/entities/post/api/query-keys.ts","./src/entities/post/api/services.ts","./src/entities/post/api/hooks/use-get-post-detail.ts","./src/entities/post/model/post-list-query-schema.ts","./src/entities/post/model/types.ts","./src/entities/post/model/hooks/use-post-list-filter-search-params.ts","./src/entities/tag/index.ts","./src/entities/tag/api/query-keys.ts","./src/entities/tag/api/services.ts","./src/entities/tag/api/hooks/use-get-tags.ts","./src/entities/tag/model/types.ts","./src/entities/user/index.ts","./src/entities/user/api/query-keys.ts","./src/entities/user/api/services.ts","./src/entities/user/api/hooks/use-get-user-detail.ts","./src/entities/user/api/hooks/use-get-users.ts","./src/entities/user/model/types.ts","./src/entities/user/ui/userdetaildialog.tsx","./src/entities/user/ui/index.ts","./src/features/add-comment/index.ts","./src/features/add-comment/api/services.ts","./src/features/add-comment/api/hooks/use-create-comment-mutation.ts","./src/features/add-comment/ui/addcommentdialog.tsx","./src/features/add-post/api/services.ts","./src/features/add-post/api/hooks/use-create-post-mutation.ts","./src/features/add-post/ui/addpostdialog.tsx","./src/features/delete-comment/index.ts","./src/features/delete-comment/api/services.ts","./src/features/delete-comment/api/hooks/use-delete-comment-mutation.ts","./src/features/delete-comment/ui/deletecommentbutton.tsx","./src/features/delete-post/api/services.ts","./src/features/delete-post/api/hooks/use-delete-post-mutation.ts","./src/features/delete-post/ui/deletepostbutton.tsx","./src/features/edit-comment/index.ts","./src/features/edit-comment/api/services.ts","./src/features/edit-comment/api/hooks/use-update-comment-mutation.ts","./src/features/edit-comment/ui/editcommentdialog.tsx","./src/features/edit-post/api/services.ts","./src/features/edit-post/api/hooks/use-update-post-mutation.ts","./src/features/edit-post/ui/editpostdialog.tsx","./src/features/like-comment/index.ts","./src/features/like-comment/api/services.ts","./src/features/like-comment/api/hooks/use-like-comment-mutation.ts","./src/features/like-comment/ui/likecommentbutton.tsx","./src/pages/post/index.tsx","./src/shared/lib/axios-instance.ts","./src/shared/lib/highlight-text.tsx","./src/shared/lib/search-params-helpler.ts","./src/shared/ui/index.ts","./src/shared/ui/button/button.tsx","./src/shared/ui/button/index.ts","./src/shared/ui/card/card.tsx","./src/shared/ui/card/index.ts","./src/shared/ui/dialog/dialog.tsx","./src/shared/ui/dialog/index.ts","./src/shared/ui/footer/footer.tsx","./src/shared/ui/footer/index.ts","./src/shared/ui/header/header.tsx","./src/shared/ui/header/index.ts","./src/shared/ui/input/input.tsx","./src/shared/ui/input/index.ts","./src/shared/ui/select/select.tsx","./src/shared/ui/select/index.ts","./src/shared/ui/table/table.tsx","./src/shared/ui/table/index.ts","./src/shared/ui/textarea/textarea.tsx","./src/shared/ui/textarea/index.ts","./src/widgets/comment-list/ui/commentssection.tsx","./src/widgets/post-dashboard/api/hooks/use-get-post-list.ts","./src/widgets/post-dashboard/ui/postdetaildialog.tsx","./src/widgets/post-dashboard/ui/postfilter.tsx","./src/widgets/post-dashboard/ui/postlist.tsx","./src/widgets/post-dashboard/ui/postpagination.tsx","./src/widgets/post-dashboard/ui/index.ts","./vite-env.d.ts"],"errors":true,"version":"5.9.2"}
